updating to 1.3
git-svn-id: https://plugins.svn.wordpress.org/absolute-privacy/trunk@222158 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ ap_mod_email.php
@@ -1,52 +1,54 @@
 <?php
-	if (function_exists('current_user_can')) {
-		if (!current_user_can('manage_options')) wp_die('You are not able to do that');
-	} else {
-		global $user_level;
-		get_currentuserinfo();
-		if ($user_level < 10) wp_die('You are not able to do that');
-	}
+
+if(!function_exists('add_action') ) exit(); //prevent direct loading
+
+if (function_exists('current_user_can')) {
+	if (!current_user_can('manage_options')) wp_die('You are not able to do that');
+} else {
+	global $user_level;
+	get_currentuserinfo();
+	if ($user_level < 10) wp_die('You are not able to do that');
+}
 
 $user_id = $_GET['id'];
-if(!$user_id) wp_die('No user ID provided. Please check the URL');
+if(!$user_id) wp_die('No user ID provided. Please check the URL'); //no ID found in URL
 
 $user = get_userdata($user_id);
+if(!$user) wp_die('Incorrect user ID provided. Please check the URL'); //ID not found in Database
 
-	if (isset($_POST['update_options'])) {
-		if ($_POST['update_options'] == "Delete User"){
-			if (!current_user_can('delete_user', $user_id) )
-				wp_die(__('You can&#8217;t delete that user.'));
+if (isset($_POST['update_options'])) {
+	if ($_POST['update_options'] == "Delete User"){
+		if (!current_user_can('delete_user', $user_id) )
+			wp_die(__('You can&#8217;t delete that user.'));
 
-			if($user_id == $current_user->ID) {
-				wp_die('You cannot delete yourself');
-			}
+		if($user_id == $current_user->ID) {
+			wp_die('You cannot delete yourself');
+		}
 			
-			wp_delete_user($user_id);
+		wp_delete_user($user_id);
+		
 		// Show a message to say we've done something
 		echo '<div class="updated"><p>' . __('User deleted') . '</p></div>';
+		
 		return;
-		}
-		if ($_POST['update_options'] == "Approve User"){
-					$user_role = new WP_User($user_id);
+	}
+	if ($_POST['update_options'] == "Approve User"){
+		$user_role = new WP_User($user_id);
+		$user_role->set_role("subscriber"); //change their role to subscriber
+		
+		$email = new absolutePrivacy(); 
+		$email->handleEmail($user_id, $type='account_approved'); //send approval email
 
-					$user_role->set_role("subscriber");
-
-			$headers = "MIME-Version: 1.0\n" .
-	    				"From: ". get_option('admin_email');   
-			$message = "Dear " . $user->user_firstname .",\n";
-			$message .= "Your account with ".get_bloginfo('name')." has been approved. You may login using the following info. \n";
-			$message .= "Username: " . $user->user_login . "\n";
-			$message .= "URL: " . get_bloginfo('url');
+		// Show a message to say we've done something
+		echo '<div class="updated"><p>' . __('User Approved. Notification sent via email.') . '</p></div>';
 	
-			@newuser_mail($user->user_email, 'Your Account Has Been Approved', $message, $headers);  //email the user telling them they've been approved
+		return;
+	}
+}	
+?> 
 
-			// Show a message to say we've done something
-			echo '<div class="updated"><p>' . __('User Approved. Notification sent via email.') . '</p></div>';
-			return;
-		}
-	}	
-		?> <div class="wrap">
-					<h2>Absolute Privacy: User Waiting Approval</h2>
+	<div class="wrap">
+		<h2>Absolute Privacy: User Waiting Approval</h2>
 					
 	<?php $user_role = new WP_User($user->ID);
 			$capabilities = $this->capabilities;
@@ -59,11 +61,11 @@ $user = get_userdata($user_id);
 ?>
 
     	
-<div style="float: left; width: 660px; margin: 5px;">	
+<div class="wrap" style="float: left; width: 60%; margin: 5px;">	
 		<form method="post" action="">
 
 
-<table class="widefat" cellspacing="0">
+<table class="widefat" cellspacing="0" >
 
 
 
@@ -90,7 +92,7 @@ $user = get_userdata($user_id);
 
 </div>
 
-<div style="float: left; width: 300px; margin: 5px; ">
+<div style="float: left; width: 30%; margin: 5px; ">
 <table name="pl_donate" class="widefat fixed" style="margin-bottom: 10px;" cellspacing="0">
 
 	<thead>
@@ -126,8 +128,7 @@ $user = get_userdata($user_id);
 			<td>
 				<ul style="font-size: 1.0em;">
 					<li><a href="http://www.johnkolbert.com/portfolio/wp-plugins/absolute-privacy/" title="Go to Plugin Homepage">Plugin Homepage</a></li>
-					<li><a href="http://www.johnkolbert.com/forum/absolute-privacy/" title="Go to Support Forum">Plugin Support Forum</a></li>
-					<li><a href="http://www.johnkolbert.com/contact/estimate/" title="Hire Me!">Hire me to customize this plugin</a></li>
+					<li><a href="http://www.mammothapps.com/contact" title="Hire Me!">Hire me to customize this plugin</a></li>
 					
 				</ul>
 			</td>
@@ -148,7 +149,7 @@ $user = get_userdata($user_id);
 		<tr>
 			<td>
 				<p style="text-align: center; font-size: 1.2em;">Plugin created by <a href="http://www.johnkolbert.com/" title="John Kolbert WordPress Consulting">John Kolbert</a><br />
-				<span style="font-size: 0.8em;">Need Help? <a href="http://www.johnkolbert.com/contact/estimate/" title="Hire Me">Hire me.</a><br />
+				<span style="font-size: 0.8em;">Need Help? <a href="http://www.mammothapps.com/contact" title="Hire Me">Hire me.</a><br />
 				<a href="http://www.twitter.com/johnkolbert" title="Follow Me!">Follow me on Twitter!</a><br /></span>
 				</p>
 			</td>
@@ -158,7 +159,3 @@ $user = get_userdata($user_id);
 
 </div>
 </div>
-
-<?php
-
-?>
